mac80211-qca: cfg80211: update ssid during update AP operation
Currently, the AP update operation is not correctly updating new SSID.
This issue arises because the NL80211_CMD_UPDATE_AP command does not
process its NL80211_ATTR_SSID attribute. Resolve this issue by ensuring
that the SSID is updated if it is provided with the
NL80211_CMD_UPDATE_AP command. So fixing it by a patch provided by the
Qualcomm's customer support system.

Fixes: PPW-15

Signed-off-by: Marvin Lu <marvin.lu@wnc.com.tw>