hostapd: fix crash during non-MLO to MLO config reload
Currently, when an Access Point (AP) switches from a non-ML
configuration to an ML configuration, it causes a hostapd crash.

This occurs because the memory allocation for the 'mld'
structure only happens during the AP's startup if the 'mld_ap' option
is enabled in the configuration.

Consequently, when hostapd transitions from a non-ML to an ML
configuration, the MLD structure's memory remains unallocated, and
attempting to access this unassigned pointer leads to a crash.

To resolve this issue, initiate hostapd_init() whenever the AP
transitions from ML to non-ML configuration, or vice versa. This
action will ensure that the memory allocation for the MLD structure is
properly executed.

QCA case: 07364813

Fixes: PPM-3061

Signed-off-by: Houssem Dafdouf <houssem.dafdouf_ext@softathome.com>