ipq95xx: add hostapd-qca dependency
For Freedom ATH12.4.0 build, the hostapd-qca dependency is added in
prplmesh package originally. The dependencies for package hostapd-qca
are hostapd-qca-openssl wpa-supplicant-qca-openssl.

However, the dependency is changed to wpad-qca-mesh-openssl in Freedom
ATH13.0. To allow same prplmesh package with Freedom ATH12.4.0
and ATH13.0, move these dependencies definition for Freedom to ipq95xx.

PCF-1798

Signed-off-by: Marvin Lu <marvin.lu@wnc.com.tw>