hostapd: Fix crash when handling AP-DISABLED command
When the INTERFACE-DISABLED command is issued, it brings down
the Wi-Fi interface. If the AP-DISABLED command is subsequently
sent by PWHM, a crash may occur because the VAP has already been
cleaned up.
This update adds a safeguard in hostapd_mld_interface_freed() to
check if hapd->mld is empty. If so, the function exits early to
prevent further processing and avoid the crash.

Signed-off-by: Ram Chandra Jangir <rjangir@qti.qualcomm.com>