PROVISORY: ipq95xx: require spatch from coccinelle for mac80211 build
The builds for Freedom using ATH13.1ED BSP are currently failing with
following error:

  bash: line 1: spatch: command not found

This is happening as the mac80211 package is modifying the source code
tree another time to amend the kzalloc/kmalloc usage for memory
debugging use case.

It should be handled with existing tools, it is currently being
addressed internally, but in the meantime we're going to demand
spatch/coccinelle for the build.

References: PCF-2142
Link: https://gitlab.com/prpl-foundation/prplos/prplos/-/jobs/11783721716#L9686
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>