mac80211: nl80211: fix update ssid during update AP operation
Currently, the AP update operation is not correctly updating new SSID.
This issue arises because the NL80211_CMD_UPDATE_AP command does not
process its NL80211_ATTR_SSID attribute. Resolve this issue by ensuring
that the SSID is updated if it is provided with the
NL80211_CMD_UPDATE_AP command.

This fix was originally made by Nikhil Gurram in branch dev/ath13.1.

Fixes: PPW-1127

Signed-off-by: Nikhil Gurram <nikhgurr@qti.qualcomm.com>
Signed-off-by: Marvin Lu <marvin.lu@wnc.com.tw>