prpl: arm: ipq9574: fix bug for Wi-Fi 6GHz not work when enabling PCIE0 for 5G modem
Add a DTS property to support 16 MSIs for four-slot PCIe.
To support four-slot PCIe, the number of MSI interrupts is changed to 16 for PCIe0 and PCIe1.

Currently 96 MSI interrupts has been used for three slot
PCI. To support four slot PCI, number of MSI has to be split
in below order

PCIe0 - 16,
PCIe1 - 16,
PCIe2 - 32,
PCIe3 - 32

Made changes to configure the MSI dynamically based on the DTS
property in PCI node. Based on the number of MSI corresponding
ring mask will be taken for DP grouping.

Signed-off-by: Tim Hsu <tim.hsu@wnc.com.tw>